Refactor listeners
